backend/requirements.txt
fastapi == 0.115.12
fastapi == 0.115.12
uvicorn == 0.34.0
requests == 2.32.3
python-dotenv == 1.1.0
langchain-community == 0.3.20
langchain-openai == 0.3.9
langchain-text-splitters == 0.3.7
chromadb == 0.6.3
numpy >=1.26.2